+++ indexcov/indexcov.go
@@ -2,6 +2,7 @@ package indexcov
 
 import (
 	"fmt"
+	"html/template"
 	"io"
 	"log"
 	"os"
@@ -18,12 +19,16 @@ import (
 	"github.com/brentp/go-chartjs/types"
 )
 
+// Ploidy indicates the expected ploidy of the samples.
+var Ploidy = 2
+
 var cli = struct {
 	Prefix    string   `arg:"-p,required,help:prefix for output files"`
 	IncludeGL bool     `arg:"-e,help:plot GL chromosomes like: GL000201.1 which are not plotted by default"`
+	Sex       []string `arg:"-X,help:name of the sex chromosome(s) used to infer sex; The first will be used to populate the sex column in a ped file."`
 	Chrom     string   `arg:"-c,help:optional chromosome to extract depth. default is entire genome."`
 	Bam       []string `arg:"positional,required,help:bam(s) for which to estimate coverage"`
-}{}
+}{Sex: []string{"X", "Y"}}
 
 func getRefs(idx *bam.Index) []RefIndex {
 	refs := reflect.ValueOf(*idx).FieldByName("idx").FieldByName("Refs")
@@ -179,7 +184,7 @@ func getShortName(b string) string {
 }
 
 func getWriter(prefix string) (*bgzf.Writer, error) {
-	fh, err := os.Create(fmt.Sprintf("%s.indexcov.bed.gz", prefix))
+	fh, err := os.Create(fmt.Sprintf("%s-indexcov.bed.gz", prefix))
 	if err != nil {
 		return nil, err
 	}
@@ -197,6 +202,9 @@ func Main() {
 
 	chartjs.XFloatFormat = "%.0f"
 	arg.MustParse(&cli)
+	if strings.HasSuffix(cli.Prefix, "/") {
+		cli.Prefix = cli.Prefix + "qc"
+	}
 
 	rdr, err := os.Open(cli.Bam[0])
 	if err != nil {
@@ -247,13 +255,14 @@ func Main() {
 		panic(err)
 	}
 
-	rfh, err := os.Create(fmt.Sprintf("%s.indexcov.roc", cli.Prefix))
+	rfh, err := os.Create(fmt.Sprintf("%s-indexcov.roc", cli.Prefix))
 	if err != nil {
 		panic(err)
 	}
 
 	// we plot all the coverage roc charts in a single html file.
 	charts := make([]chartjs.Chart, 0, len(refs))
+	sexes := make(map[string][]float64)
 
 	fmt.Fprintf(bgz, "#chrom\tstart\tend\t%s\n", strings.Join(names, "\t"))
 	for ir, ref := range refs {
@@ -274,6 +283,13 @@ func Main() {
 
 			CountsAtDepth(depths[k], counts[k])
 		}
+
+		for _, x := range cli.Sex {
+			if x == chrom && len(depths[longesti]) > 0 {
+				sexes[chrom] = GetCN(depths)
+			}
+		}
+
 		for i := 0; i < len(depths[longesti]); i++ {
 			fmt.Fprintf(bgz, "%s\t%d\t%d\t%s\n", chrom, i*16384, (i+1)*16384, depthsFor(depths, i))
 		}
@@ -291,17 +307,81 @@ func Main() {
 		}
 	}
 	chartjs.XFloatFormat = "%.2f"
-	plotRocs(cli.Prefix, charts)
+	saveCharts(fmt.Sprintf("%s-indexcov-roc.html", cli.Prefix), "", charts...)
 	bgz.Close()
+	writePed(sexes, cli.Sex, names, cli.Prefix)
 }
 
-func plotRocs(prefix string, charts []chartjs.Chart) {
-	wtr, err := os.Create(fmt.Sprintf("%s-indexcov-roc.html", prefix))
+func writePed(sexes map[string][]float64, keys []string, samples []string, prefix string) {
+	if len(sexes) == 0 {
+		return
+	}
+	for _, k := range keys {
+		if _, ok := sexes[k]; !ok {
+			fmt.Printf("chromosome %s not found. not writing ped\n")
+			os.Exit(1)
+		}
+	}
+	sexes["_inferred"] = make([]float64, len(sexes[keys[0]]))
+	f, err := os.Create(fmt.Sprintf("%s-indexcov.ped", prefix))
+	if err != nil {
+		panic(err)
+	}
+	hdr := make([]string, len(keys))
+	for i, k := range keys {
+		hdr[i] = "CN" + k
+	}
+	fmt.Fprintf(f, "#family_id\tsample_id\tpaternal_id\tmaternal_id\tsex\tphenotype\t%s\n", strings.Join(hdr, "\t"))
+
+	tmpl := "unknown\t%s\t-9\t-9\t%d\t-9\t"
+	for i, s := range samples {
+		inferred := int(0.5 + sexes[keys[0]][i])
+		fmt.Fprintf(f, tmpl, s, inferred)
+		sexes["_inferred"][i] = float64(inferred)
+		s := make([]string, 0, len(keys))
+		for _, k := range keys {
+			s = append(s, fmt.Sprintf("%.2f", sexes[k][i]))
+		}
+		fmt.Fprintln(f, strings.Join(s, "\t"))
+	}
+	if len(keys) > 1 {
+		chart, customjs, err := plotSex(sexes, keys[:2], samples)
+		if err != nil {
+			panic(err)
+		}
+		saveCharts(fmt.Sprintf("%s-indexcov-sex.html", cli.Prefix), customjs, chart)
+	}
+}
+
+// GetCN returns an float per sample estimating the number of copies of that chromosome.
+// It is a very crude estimate, but that's what indexcov is and it tends to work well.
+func GetCN(depths [][]float32) []float64 {
+	if depths == nil {
+		return nil
+	}
+	meds := make([]float64, 0, len(depths))
+	for _, d := range depths {
+		tmp := make([]float32, 0, len(d))
+		for _, dp := range d {
+			// exclude sites that are exactly 0 as these are the centromere.
+			if dp != 0 {
+				tmp = append(tmp, dp)
+			}
+		}
+		sort.Slice(tmp, func(i, j int) bool { return tmp[i] < tmp[j] })
+		med := float64(float32(Ploidy) * tmp[int(float64(len(tmp))*0.5)])
+		meds = append(meds, med)
+	}
+	return meds
+}
+
+func saveCharts(path string, customjs string, charts ...chartjs.Chart) {
+	wtr, err := os.Create(path)
 	if err != nil {
 		panic(err)
 	}
 	defer wtr.Close()
-	if err := chartjs.SaveCharts(wtr, map[string]interface{}{"height": 800, "width": 800}, charts...); err != nil {
+	if err := chartjs.SaveCharts(wtr, map[string]interface{}{"height": 800, "width": 800, "custom": template.JS(customjs)}, charts...); err != nil {
 		panic(err)
 	}
 }