+++ HISTORY.md
@@ -3,6 +3,7 @@ v0.1.9 (dev)
 + `indexcov`: fix for sparse bams and other edge-cases.
 + `indexcov`: performance improvements.
 + `indexcov`: output html with plot of coverage.
++ `indexcov`: plot and report sex inferred from coverage on allosomes.
 
 v0.1.8
 ======